backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.14.2
billiard==4.2.2
cattrs==25.3.0
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect==1.3.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
fake-useragent==2.2.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
idna==3.11
jieba3k==0.35.1
joblib==1.5.2
kombu==5.5.4
lxml==6.0.2
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pgvector==0.4.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
redis==6.4.0
regex==2025.9.18
requests==2.32.5
requests-cache==1.2.1
requests-file==3.0.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tenacity==9.1.2
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
openai>=1.0.0
tiktoken>=0.5.0

# Production server and deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
